+++ src/components/navigation/Navbar.jsx
@@ -1,13 +1,7 @@
-import { Link } from 'react-router-dom';
-
-import Logout from './logout';
 import MenuIcon from './MenuIcon';
 import SearchIcon from './SearchIcon';
-
-const links = [
-  { path: '/', text: 'Home' },
-  { path: '/dummy', text: 'Dummy' },
-];
+import DesktopLinks from './DesktopLinks';
+import MobileMenu from './MobileMenu';
 
 const handleMenuClick = () => {
   const menu = document.querySelector('#menu');
@@ -21,30 +15,9 @@ function Navbar() {
 
       <span className="hidden lg:inline-block">Logo</span>
 
-      <div
-        className="absolute top-0 right-0 w-full text-left bg-gray-100 p-4 lg:hidden hidden"
-        id="menu"
-      >
-        <MenuIcon onClick={handleMenuClick} />
+      <MobileMenu />
 
-        <ul className="text-center">
-          {links.map((link) => (
-            <li
-              key={link.path}
-              className="py-2 hover:bg-gray-200 rounded"
-              onClick={handleMenuClick}
-            >
-              <Link to={link.path}>{link.text}</Link>
-            </li>
-          ))}
-          <li
-            className="py-2 hover:bg-gray-200 rounded"
-            onClick={handleMenuClick}
-          >
-            <Logout />
-          </li>
-        </ul>
-      </div>
+      <DesktopLinks />
 
       <SearchIcon />
     </nav>
